them product va category

+++ src/main/java/iuh/fit/se/services/ProductService.java
@@ -1,25 +1,32 @@
 package iuh.fit.se.services;
 
-
 import iuh.fit.se.entities.Product;
 import iuh.fit.se.utils.ApiResponse;
 import iuh.fit.se.utils.PageResponse;
 
 public interface ProductService {
 
-    ApiResponse findById(Long id);
+    // Lấy sản phẩm theo id
+    ApiResponse findById(int id);
 
+    // Lấy tất cả sản phẩm
     ApiResponse findAll();
 
+    // Lấy danh sách sản phẩm có phân trang
     PageResponse<Product> findAllWithPaging(int page, int size, String sort);
 
+    // Thêm mới sản phẩm
     ApiResponse save(Product product);
 
-    ApiResponse update(Long id, Product product);
+    // Cập nhật sản phẩm
+    ApiResponse update(int id, Product product);
 
-    ApiResponse delete(Long id);
+    // Xóa sản phẩm
+    ApiResponse delete(int id);
 
+    // Tìm kiếm sản phẩm theo từ khóa
     ApiResponse search(String keyword);
 
-    ApiResponse findByCategory(Long categoryId);
+    ApiResponse findByCategory(int categoryId);
+
 }
